--- src/main/java/org/sopt/carena/member/appliacation/port/in/OAuth2LoginUseCase.java
@@ -1,8 +1,0 @@
-package org.sopt.carena.member.appliacation.port.in;
-
-import org.sopt.carena.member.appliacation.dto.command.OAuth2LoginCommand;
-import org.sopt.carena.member.appliacation.dto.view.OAuth2LoginResult;
-
-public interface OAuth2LoginUseCase {
-    OAuth2LoginResult processLogin(OAuth2LoginCommand command);
-}
