--- src/main/java/org/sopt/carena/member/appliacation/port/in/RefreshTokenUseCase.java
@@ -1,7 +1,0 @@
-package org.sopt.carena.member.appliacation.port.in;
-
-import org.sopt.carena.member.appliacation.dto.view.TokenRefreshView;
-
-public interface RefreshTokenUseCase {
-    TokenRefreshView refreshAccessToken(String refreshToken);
-}
