--- src/main/java/org/sopt/carena/member/appliacation/port/in/SignupUseCase.java
@@ -1,9 +1,0 @@
-package org.sopt.carena.member.appliacation.port.in;
-
-import org.sopt.carena.member.appliacation.dto.command.SignUpCommand;
-import org.sopt.carena.member.appliacation.dto.view.SignupView;
-
-public interface SignupUseCase {
-    SignupView signup(SignUpCommand command);
-}
-
