+++ src/main/java/org/sopt/carena/member/application/dto/command/OAuth2LoginCommand.java
@@ -1,4 +1,4 @@
-package org.sopt.carena.member.appliacation.dto.command;
+package org.sopt.carena.member.application.dto.command;
 
 import org.sopt.carena.member.domain.AuthType;
