+++ src/main/java/org/sopt/carena/member/application/dto/view/NewMemberSignupView.java
@@ -1,4 +1,4 @@
-package org.sopt.carena.member.appliacation.dto.view;
+package org.sopt.carena.member.application.dto.view;
 
 public record NewMemberSignupView(
         String tempToken
