+++ src/main/java/org/sopt/carena/member/application/dto/view/OAuth2LoginResult.java
@@ -1,4 +1,4 @@
-package org.sopt.carena.member.appliacation.dto.view;
+package org.sopt.carena.member.application.dto.view;
 
 public interface OAuth2LoginResult {
     default boolean needsSignup() {
