+++ src/main/java/org/sopt/carena/member/application/dto/view/SignupView.java
@@ -1,4 +1,4 @@
-package org.sopt.carena.member.appliacation.dto.view;
+package org.sopt.carena.member.application.dto.view;
 
 public record SignupView(
         String accessToken,
