+++ src/main/java/org/sopt/carena/member/application/port/in/SignupUseCase.java
@@ -1,0 +1,9 @@
+package org.sopt.carena.member.application.port.in;
+
+import org.sopt.carena.member.application.dto.command.SignUpCommand;
+import org.sopt.carena.member.application.dto.view.SignupView;
+
+public interface SignupUseCase {
+    SignupView signup(SignUpCommand command);
+}
+
