+++ src/main/java/org/sopt/carena/member/application/service/OAuth2LoginService.java
@@ -1,13 +1,13 @@
-package org.sopt.carena.member.appliacation.service;
+package org.sopt.carena.member.application.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.sopt.carena.member.appliacation.dto.command.OAuth2LoginCommand;
-import org.sopt.carena.member.appliacation.dto.view.*;
-import org.sopt.carena.member.appliacation.port.in.OAuth2LoginUseCase;
-import org.sopt.carena.member.appliacation.port.out.JoinTokenStore;
-import org.sopt.carena.member.appliacation.port.out.MemberPersistencePort;
-import org.sopt.carena.member.appliacation.port.out.RefreshTokenStore;
-import org.sopt.carena.member.appliacation.service.util.JwtTokenGenerator;
+import org.sopt.carena.member.application.dto.command.OAuth2LoginCommand;
+import org.sopt.carena.member.application.dto.view.*;
+import org.sopt.carena.member.application.port.in.OAuth2LoginUseCase;
+import org.sopt.carena.member.application.port.out.JoinTokenStore;
+import org.sopt.carena.member.application.port.out.MemberPersistencePort;
+import org.sopt.carena.member.application.port.out.RefreshTokenStore;
+import org.sopt.carena.member.application.service.util.JwtTokenGenerator;
 import org.sopt.carena.member.domain.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
